+++ RealmServer/Handlers/OnTutorialFlag.cs
@@ -1,0 +1,11 @@
+﻿using RealmServer.PacketReader;
+
+namespace RealmServer.Handlers
+{
+    public class OnTutorialFlag
+    {
+        public static void Handler(RealmServerSession session, CMSG_TUTORIAL_FLAG handler)
+        {
+        }
+    }
+}