+++ RealmServer/Handlers/OnTutorialReset.cs
@@ -1,0 +1,15 @@
+﻿namespace RealmServer.Handlers
+{
+    public class OnTutorialReset
+    {
+        public static void Handler(RealmServerSession session, byte[] data)
+        {
+            /*
+            Dim i As Integer
+            For i = 0 To 31
+            Client.Character.TutorialFlags(i) = 0
+            Next
+            */
+        }
+    }
+}