+++ RealmServer/PacketReader/CMSG_TUTORIAL_FLAG.cs
@@ -1,0 +1,12 @@
+﻿namespace RealmServer.PacketReader
+{
+    public sealed class CMSG_TUTORIAL_FLAG : Common.Network.PacketReader
+    {
+        public int Flag;
+
+        public CMSG_TUTORIAL_FLAG(byte[] data) : base(data)
+        {
+            Flag = ReadInt32();
+        }
+    }
+}