+++ RealmServer/PacketServer/SMSG_TUTORIAL_FLAGS.cs
@@ -7,7 +7,7 @@ namespace RealmServer.PacketServer
         public SMSG_TUTORIAL_FLAGS() : base(RealmEnums.SMSG_TUTORIAL_FLAGS)
         {
             for (var i = 0; i < 8; i++)
-                Write(-1);
+                Write(0);
         }
     }
 }